requirements.txt
Django==2.1
djangorestframework==3.8.2
pytz==2018.5
mysqlclient==1.3.13
python-dotenv==0.9.1